Fixing flaky tests
Verification for operate permissions in PipelineVisibility triggered off the schedule API to test if the user had the required permissions or not. This caused some of these tests to be flaky if the pipeline run did not finish in time before the next set of assertions around rerunning stages. Making the test wait for the run to finish before moving on to next set of steps